wip - pause to do smaller refactorings
